modify all components to use recoil
