+++ src/DataTable.js
@@ -20,9 +20,6 @@ import ViewColumn from "@material-ui/icons/ViewColumn";
 import { cloneDeep } from "lodash";
 import { appState } from "./state.js";
 import {
-  RecoilRoot,
-  atom,
-  selector,
   useRecoilState,
   useRecoilValue,
 } from 'recoil';
@@ -65,6 +62,7 @@ function DataTable(args) {
   const objState = useRecoilValue(appState[args.id]);
   const [queryState, setQueryState] = useRecoilState(appState[ objState.dataref ]);
 
+/*
   const runUrlFetchQuery = (qid) => {
     //console.log("Run Url Fetch Query: " + qid);
     const url = queryState.query;
@@ -78,7 +76,7 @@ function DataTable(args) {
         setQueryState(copyState);
       });
   };
-
+*/
   // initialize the data that depends on initial fetch query
   React.useEffect(() => {
     if (queryState.backend === "urlfetch" && queryState.fetch_on_init){
