+++ src/QueryForm.js
@@ -5,9 +5,9 @@ function QueryForm(args) {
   const classes = useStyles();
   return (
     <form className={classes.root} noValidate autoComplete="off">
-      {args.widgets.map((spec) =>
-        (<Widget wspec={spec} callbacks={args.callbacks} state={args.state} widgets={args.widget}/>)
-      )} 
+      {args.widgets.map((spec) => (
+        <Widget wspec={spec} widgets={args.widget} />
+      ))}
     </form>
   );
 }
