+++ src/TabContainer.js
@@ -1,20 +1,16 @@
 // for tab containers
-import PropTypes from 'prop-types';
-import AppBar from '@material-ui/core/AppBar';
-import Tabs from '@material-ui/core/Tabs';
-import Tab from '@material-ui/core/Tab';
-import Typography from '@material-ui/core/Typography';
-import Box from '@material-ui/core/Box';
+import PropTypes from "prop-types";
+import AppBar from "@material-ui/core/AppBar";
+import Tabs from "@material-ui/core/Tabs";
+import Tab from "@material-ui/core/Tab";
+import Typography from "@material-ui/core/Typography";
+import Box from "@material-ui/core/Box";
 
 import { useStyles } from "./Styles.js";
 import { Widget } from "./Widget.js";
 import { appState } from "./state.js";
 import { cloneDeep } from "lodash";
-import {
-  useRecoilState,
-  useRecoilValue,
-} from 'recoil';
-
+import { useRecoilState } from "recoil";
 
 function TabPanel(props) {
   const { children, value, index, ...other } = props;
@@ -29,7 +25,7 @@ function TabPanel(props) {
     >
       {value === index && (
         <Box p={3}>
-          <Typography component={'span'} >{children}</Typography>
+          <Typography component={"span"}>{children}</Typography>
         </Box>
       )}
     </div>
@@ -45,19 +41,14 @@ TabPanel.propTypes = {
 function a11yProps(index) {
   return {
     id: `simple-tab-${index}`,
-    'aria-controls': `simple-tabpanel-${index}`,
+    "aria-controls": `simple-tabpanel-${index}`,
   };
 }
 
 function TabContainer(args) {
   const classes = useStyles();
-//  const [value, setValue] = React.useState(0);
-
-//  const handleChange = (event, newValue) => {
-//    setValue(newValue);
-//  };
-  console.log("tabcontainer args.id = " + args.id);
-  console.log(appState);
+  //console.log("tabcontainer args.id = " + args.id);
+  //console.log(appState);
 
   const [objState, setObjState] = useRecoilState(appState[args.id]);
 
@@ -70,23 +61,42 @@ function TabContainer(args) {
     setObjState(copyState);
   };
   return (
-    <div className={classes.root} >
-      <AppBar key={"appbar-"+args.id} position="static" style={{ minWidth: "100%" }}>
-        <Tabs key={"tabs-"+args.id} value={objState.value} onChange={handleTabChange} 
-            aria-label="simple tabs example"
-            variant="scrollable"
-            scrollButtons="auto"
-            className={classes.customTwo}
-            >
-          { args.tabs.map( (tab) =>  
-                (<Tab key={"tab-" + tab.id} label={tab.label} {...a11yProps(tab.idx)} />)) }
+    <div className={classes.root}>
+      <AppBar
+        key={"appbar-" + args.id}
+        position="static"
+        style={{ minWidth: "100%" }}
+      >
+        <Tabs
+          key={"tabs-" + args.id}
+          value={objState.value}
+          onChange={handleTabChange}
+          aria-label="simple tabs example"
+          variant="scrollable"
+          scrollButtons="auto"
+          className={classes.customTwo}
+        >
+          {args.tabs.map((tab) => (
+            <Tab
+              key={"tab-" + tab.id}
+              label={tab.label}
+              {...a11yProps(tab.idx)}
+            />
+          ))}
         </Tabs>
       </AppBar>
-      { args.tabs.map( (tab) =>  
-                (<TabPanel key={"tabpanel-"+tab.id} value={objState.value} index={tab.idx} style={{ minWidth: "100%" }}>
-                    { tab.widgets.map( (spec) => (<Widget wspec={spec} />) ) }
-                </TabPanel>
-                )) }
+      {args.tabs.map((tab) => (
+        <TabPanel
+          key={"tabpanel-" + tab.id}
+          value={objState.value}
+          index={tab.idx}
+          style={{ minWidth: "100%" }}
+        >
+          {tab.widgets.map((spec) => (
+            <Widget wspec={spec} />
+          ))}
+        </TabPanel>
+      ))}
     </div>
   );
 }
