+++ src/Widget.js
@@ -1,27 +1,14 @@
-import Grid from '@material-ui/core/Grid';
-import Paper from '@material-ui/core/Paper';
+import Grid from "@material-ui/core/Grid";
+//import Paper from "@material-ui/core/Paper";
 import { TabContainer } from "./TabContainer.js";
 import { DataTable } from "./DataTable.js";
 import { QueryForm } from "./QueryForm.js";
-import { useStyles } from "./Styles.js";
-import { appState } from "./state.js";
-
-import TextField from "@material-ui/core/TextField";
-import InputLabel from "@material-ui/core/InputLabel";
-import MenuItem from "@material-ui/core/MenuItem";
-//import FormHelperText from "@material-ui/core/FormHelperText";
-import FormControl from "@material-ui/core/FormControl";
-import Select from "@material-ui/core/Select";
-import Button from "@material-ui/core/Button";
-
-import {
-  useRecoilState,
-  useRecoilValue,
-} from 'recoil';
+import { TextInput } from "./TextInput.js";
+import { Menu } from "./Menu.js";
+import { MyButton } from "./MyButton.js";
 
 // Grid will use 12 units has maxwidth
 function Widget(args) {
-  const classes = useStyles();
 
   switch (args.wspec.type) {
     case "tabcontainer":
@@ -44,98 +31,77 @@ function Widget(args) {
           />
         </Grid>
       );
-/*
     case "form":
       return (
         <Grid item xs={12}>
-          <QueryForm
-            key={args.wspec.id}
-            widgets={args.wspec.widgets}
-          />
+          <QueryForm key={args.wspec.id} widgets={args.wspec.widgets} />
         </Grid>
       );
 
     case "text_input":
       return (
-        <TextField
+        <TextInput
           key={args.wspec.id}
           id={args.wspec.id}
           label={args.wspec.label}
-          onChange={(event) => args.callbacks.handleTextChange(args.wspec.id, event)}
-          variant="outlined"
         />
       );
     case "menu":
       return (
-        <FormControl
-          key={args.wspec.id}
-          variant="filled"
-          className={classes.formControl}
-        >
-          <InputLabel id="demo-simple-select-filled-label">
-            {args.wspec.label}
-          </InputLabel>
-          <Select
-            labelId="demo-simple-select-filled-label"
-            id="qtypedropdown"
-            value={args.state[args.wspec.id].value}
-            onChange={(event) => args.callbacks.handleSelectChange(args.wspec.id, event)}
-          >
-            <MenuItem value="">
-              <em>None</em>
-            </MenuItem>
-            {args.wspec.values.map((v) => {
-              return (
-                <MenuItem key={v.id} value={v.value}>
-                  {v.display}
-                </MenuItem>
-              );
-            })}
-          </Select>
-        </FormControl>
+        <Menu
+          id={args.wspec.id}
+          label={args.wspec.label}
+          values={args.wspec.values}
+        />
       );
     case "button":
       return (
-        <Button
+        <MyButton
           key={args.wspec.id}
           id={args.wspec.id}
-          trigger={args.wspec.trigger}
-          variant="contained"
-          onClick={() => args.callbacks.handleClick(args.wspec.id)}
-        >
-          {args.wspec.label}
-        </Button>
-      );
-*/
+          label={args.wspec.label}
+        />);
     case "image":
       return (
-        <Grid container xs={args.wspec.width} 
-            justifyContent={args.wspec.justify}
-            alignItems="center"
-            >
+        <Grid
+          container
+          xs={args.wspec.width}
+          justifyContent={args.wspec.justify}
+          alignItems="center"
+        >
           <Grid item xs={args.wspec.width}>
-            <img alt="portrait" border="1" style={{
-                "marginTop": "2em", 
-                    height: "220px"}} 
-                src={args.wspec.value}/>
-          </Grid >
-        </Grid >
+            <img
+              alt="portrait"
+              border="1"
+              style={{
+                marginTop: "2em",
+                height: "220px",
+              }}
+              src={args.wspec.value}
+            />
+          </Grid>
+        </Grid>
       );
     case "text":
       return (
-        <Grid container xs={args.wspec.width} 
-            justifyContent={args.wspec.justify}
-            >
+        <Grid
+          container
+          xs={args.wspec.width}
+          justifyContent={args.wspec.justify}
+        >
           <Grid item xs={args.wspec.width}>
-            <div key={args.wspec.id} dangerouslySetInnerHTML={{__html: args.wspec.value}}/>
-          </Grid >
-        </Grid >
+            <div
+              key={args.wspec.id}
+              dangerouslySetInnerHTML={{ __html: args.wspec.value }}
+            />
+          </Grid>
+        </Grid>
       );
-//          <Paper className={classes.paper}>
-//          </Paper>
+    //          <Paper className={classes.paper}>
+    //          </Paper>
     default:
       return null;
   }
 }
 
-export { Widget } ;
+export { Widget };
