+++ src/index.js
@@ -1,14 +1,10 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
-import { App } from './App';
-import reportWebVitals from './reportWebVitals';
-import { ApolloClient, InMemoryCache, useQuery, gql } from '@apollo/client';
-import { ApolloProvider } from '@apollo/client/react';
+import React from "react";
+import ReactDOM from "react-dom";
+import "./index.css";
+import { App } from "./App";
+import reportWebVitals from "./reportWebVitals";
 
-import {
-  RecoilRoot,
-} from 'recoil';
+import { RecoilRoot } from "recoil";
 
 import { app_spec } from "./appSpec";
 
@@ -16,11 +12,11 @@ ReactDOM.render(
   <React.StrictMode>
     <RecoilRoot>
       <React.Suspense fallback={<div>Loading...</div>}>
-      <App spec={app_spec}/>
+        <App spec={app_spec} />
       </React.Suspense>
     </RecoilRoot>
   </React.StrictMode>,
-  document.getElementById('root')
+  document.getElementById("root")
 );
 
 // If you want to start measuring performance in your app, pass a function
